% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logtest.R
\name{logtest}
\alias{logtest}
\title{Multiple curve log-rank test}
\usage{
logtest(data, cat1, cat2)
}
\arguments{
\item{data}{Data frame containing lifespan data for at least two groups.}

\item{cat1}{Group to separate data by.}

\item{cat2}{Optional second group to separate data by.}
}
\value{
A data frame of the significance level between each pair of groups.
}
\description{
Multiple curve log-rank test
}
\examples{
data <- data.frame('Group'=c('A','B','B','A','B','A','C','C','C'),
  'Lifespan'=c(2,7,8,3,6,1,15,16,18))
logtest(data, Group)
}
