% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirect.R
\name{multirect}
\alias{multirect}
\title{Calculate rectangularity of multiple survival curves}
\usage{
multirect(data, cat1, cat2, cat3)
}
\arguments{
\item{data}{Data frame containing lifespan data for at least two groups.}

\item{cat1}{Group to separate data by.}

\item{cat2}{Optional second group to separate data by.}

\item{cat3}{Optional third group to separate data by.}
}
\value{
A data frame containing rectangularity metrics for each group.
}
\description{
Calculate rectangularity of multiple survival curves
}
\examples{
data <- data.frame('Group1'=c('A','B','A','B','C','C'),'Group2'=c('c','a','b','b','c','a'),
  'Lifespan'=c(2,7,8,3,6,1))
multirect(data,Group1,Group2)
}
